Complete rewrite
I rewrote the original app from scratch, using rivets and bootstrap.  I
am not done with it yet and I still need to update what is in the app
store but it is functional.

This new iteration is meant to work offline.  It caches the data
locally and if the user is offline, it just uses the most recent cache.
 If the user is offline, it also hides the twitter modal.